config: rename androidapp package to app
Renamed the package suffix inside the app module from "androidapp" to "app" and updated all files accordingly.
